+++ src/views/List/StockItem/PurchasedStock/PurchasedStock.tsx
@@ -4,10 +4,7 @@ import styled from 'styled-components';
 
 import { updateCheckedItems } from '../../../../features/checkedItems/checkedItemsSlice';
 import { selectIsPurchasedItemChecked } from '../../../../features/checkedItems/selectors';
-import {
-  selectIsMainGroupSelected,
-  selectSelectedGroupId,
-} from '../../../../features/groups/selectors';
+import { selectIsMainGroupSelected } from '../../../../features/groups/selectors';
 import { selectPurchasedItemsById } from '../../../../features/stockList/selectors';
 import {
   updatePurchaseItem,
@@ -16,26 +13,16 @@ import {
 import { EditUserItemServiceData } from '../../../../service/userStocks/type';
 import { selectIsLoggedIn } from '../../../../features/user/selectors';
 
-import { TableCell, TableRow } from '../../../../components/Table';
+import { TableRow } from '../../../../components/Table';
 import { BorderButton } from '../../../../components/Button';
 import { TransformedValue } from '../../../../components/Input/BaseInput';
 import userStocksService from '../../../../service/userStocks/userStocks';
-import useModal from '../../hooks/useModal';
 import { CheckboxCell } from '../components';
-import { DeleteStockModal } from '../DeleteStockModal';
 import PurchasedContent from './PurchasedContent';
 import { getChangedPurchasedData } from './utils';
 import { checkNoChange } from '../utils';
-import IconButton, { EditButton, MoreButton } from '@/components/IconButton';
-import { getSoldInfoFromPurchasedInfo } from '@/features/solds/utils';
-import { addNewSold } from '@/features/solds';
-import userSoldsService from '@/service/userSolds/service';
-import getDateAndTime from '@/utils/getDateAndTime';
-import { NewSold } from '@/repository/userSolds';
-import { DropboxItem } from '@/components/Dropbox/DropboxItem';
-import AddToGroupModal from './AddToGroupModal';
-import { removePurchasedItemFromGroup } from '@/features/groups/groupsSlice';
-import userGroupsService from '@/service/userGroups/userGroups';
+import PurchasedMainGroupAction from './PurchasedMainGroupAction';
+import PurchasedOtherGroupAction from './PurchasedOtherGroupAction';
 
 export type SetChangedInputByFieldName = <
   T extends keyof ChangedPurchasedItemInputs,
@@ -59,21 +46,18 @@ export type ChangedPurchasedItemInputs = EditUserItemServiceData;
 const PurchasedStock = ({ stockId, purchasedId }: PurchasedStockProps) => {
   const dispatch = useDispatch();
 
-  const { mainInfo, purchasedItem } = useSelector(
+  const { purchasedItem } = useSelector(
     selectPurchasedItemsById(stockId, purchasedId),
   );
   const isPurchasedItemChecked = useSelector(
     selectIsPurchasedItemChecked(stockId, purchasedId),
   );
   const isMainGroupSelected = useSelector(selectIsMainGroupSelected);
-  const selectedGroupId = useSelector(selectSelectedGroupId);
   const isLoggedIn = useSelector(selectIsLoggedIn);
 
   const [isLock, setIsLock] = useState(!purchasedItem.needInit);
   const [changedInputs, setChangedInputs] =
     useState<ChangedPurchasedItemInputs>({});
-  const deleteModal = useModal();
-  const groupModal = useModal();
 
   const setChangedInputByFieldName = useCallback<SetChangedInputByFieldName>(
     (fieldName, value) => {
@@ -92,7 +76,7 @@ const PurchasedStock = ({ stockId, purchasedId }: PurchasedStockProps) => {
       }),
     );
 
-  const onToggleLock = async () => {
+  const onToggleLock = useCallback(async () => {
     if (purchasedItem.needInit)
       dispatch(updatePurchaseItemNeedInit({ stockId, purchasedId }));
     if (isLock) return setIsLock(false);
@@ -113,51 +97,7 @@ const PurchasedStock = ({ stockId, purchasedId }: PurchasedStockProps) => {
 
     setChangedInputs({});
     setIsLock(true);
-  };
-
-  const onItemSold = async () => {
-    if (isLoggedIn) {
-      const { date, time } = getDateAndTime();
-      const soldItem: NewSold = {
-        ...purchasedItem,
-        stockId: mainInfo.stockId,
-        stockName: mainInfo.stockName,
-        soldPrice: mainInfo.currentPrice,
-      };
-      const result = await userSoldsService.addNewSolds({
-        date,
-        time,
-        solds: [soldItem],
-      });
-
-      if (!result.success) {
-        alert(result.message);
-        return;
-      }
-    }
-
-    const soldInfo = getSoldInfoFromPurchasedInfo(mainInfo, purchasedItem);
-    dispatch(addNewSold({ soldInfo, stockId: mainInfo.stockId }));
-  };
-
-  const onRemoveItemFromGroup = async () => {
-    if (isLoggedIn) {
-      const result = await userGroupsService.deletePurchasedItemFromUserGroup({
-        stockId,
-        purchasedId,
-        groupId: selectedGroupId,
-      });
-
-      if (!result.success) return;
-    }
-    dispatch(
-      removePurchasedItemFromGroup({
-        groupId: selectedGroupId,
-        stockId,
-        purchasedId,
-      }),
-    );
-  };
+  }, []);
 
   useEffect(() => {
     if (isMainGroupSelected && purchasedItem.needInit)
@@ -190,66 +130,18 @@ const PurchasedStock = ({ stockId, purchasedId }: PurchasedStockProps) => {
         setChangedInputByFieldName={setChangedInputByFieldName}
       />
       {isMainGroupSelected ? (
-        <>
-          <TableCell>
-            <StyledButtonGroup>
-              <EditButton
-                isLock={isLock}
-                onClick={onToggleLock}
-                disabled={!isMainGroupSelected}
-              />
-              <MoreButton width={80} vertical='bottom' horizontal='right'>
-                <DropboxItem
-                  onClick={onItemSold}
-                  disabled={!isLock}
-                  title='To sold list'
-                >
-                  Sold
-                </DropboxItem>
-                <DropboxItem
-                  onClick={groupModal.onOpenModal}
-                  title='Group actions'
-                >
-                  Group
-                </DropboxItem>
-                <DropboxItem
-                  onClick={deleteModal.onOpenModal}
-                  title='Delete item'
-                >
-                  Delete
-                </DropboxItem>
-              </MoreButton>
-            </StyledButtonGroup>
-          </TableCell>
-          {groupModal.showModal && (
-            <AddToGroupModal
-              stockId={stockId}
-              purchasedId={purchasedId}
-              onClose={groupModal.onCloseModal}
-            />
-          )}
-          {deleteModal.showModal && (
-            <DeleteStockModal
-              type='purchase'
-              stockId={stockId}
-              purchasedId={purchasedId}
-              onClose={deleteModal.onCloseModal}
-            />
-          )}
-        </>
+        <PurchasedMainGroupAction
+          stockId={stockId}
+          purchasedId={purchasedId}
+          isLock={isLock}
+          onToggleLock={onToggleLock}
+        />
       ) : (
-        <>
-          <TableCell>
-            <RemoveFromGroupButton>
-              <IconButton
-                width={32}
-                icon='remove'
-                onClick={onRemoveItemFromGroup}
-                title='Remove from group'
-              />
-            </RemoveFromGroupButton>
-          </TableCell>
-        </>
+        <PurchasedOtherGroupAction
+          stockId={stockId}
+          purchasedId={purchasedId}
+          isLock={isLock}
+        />
       )}
     </StyledPurchasedStock>
   );
@@ -270,15 +162,4 @@ const StyledPurchasedStock = styled(TableRow)`
       background: ${({ theme }) => theme.colors.grey400};
     }
   }
-`;
-
-export const StyledButtonGroup = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-
-const RemoveFromGroupButton = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
 `;
